+++ src/credentials.rs
@@ -10,7 +10,9 @@ use std::fs::File;
 use std::io::prelude::*;
 use std::sync::Arc;
 
-use super::jwt::{create_jwt_encoded, download_google_jwks, verify_access_token, JWKSetDTO, JWT_AUDIENCE_IDENTITY};
+use super::jwt::{
+    create_jwt_encoded, download_google_jwks, verify_access_token, JWKSetDTO, JWT_AUDIENCE_IDENTITY,
+};
 use crate::errors::FirebaseError;
 use crate::oauth2;
 
@@ -68,8 +70,9 @@ pub fn pem_to_der(pem_file_contents: &str) -> Result<Vec<u8>, Error> {
     }
 
     let base64_body = pem_file_contents.unwrap().replace("\n", "");
-    Ok(decode(&base64_body)
-        .map_err(|_| FirebaseError::Generic("Invalid private key in credentials file. Expected Base64 data."))?)
+    Ok(decode(&base64_body).map_err(|_| {
+        FirebaseError::Generic("Invalid private key in credentials file. Expected Base64 data.")
+    })?)
 }
 
 #[test]
@@ -81,11 +84,12 @@ Amtz4dJQ1YlGi0/BGhK2lg==
 -----END PRIVATE KEY-----
 "#;
     const EXPECTED: [u8; 112] = [
-        48, 130, 4, 188, 2, 1, 0, 48, 13, 6, 9, 42, 134, 72, 134, 247, 13, 1, 1, 1, 5, 0, 4, 130, 4, 166, 48, 130, 4,
-        162, 2, 1, 0, 2, 130, 1, 1, 0, 147, 110, 223, 81, 179, 105, 226, 200, 132, 68, 20, 135, 107, 132, 131, 123,
-        231, 183, 170, 255, 84, 114, 253, 88, 89, 0, 176, 87, 2, 247, 160, 250, 91, 126, 186, 47, 253, 16, 123, 88, 60,
-        145, 245, 3, 211, 114, 200, 12, 2, 134, 205, 20, 9, 21, 170, 146, 65, 40, 2, 107, 115, 225, 210, 80, 213, 137,
-        70, 139, 79, 193, 26, 18, 182, 150,
+        48, 130, 4, 188, 2, 1, 0, 48, 13, 6, 9, 42, 134, 72, 134, 247, 13, 1, 1, 1, 5, 0, 4, 130,
+        4, 166, 48, 130, 4, 162, 2, 1, 0, 2, 130, 1, 1, 0, 147, 110, 223, 81, 179, 105, 226, 200,
+        132, 68, 20, 135, 107, 132, 131, 123, 231, 183, 170, 255, 84, 114, 253, 88, 89, 0, 176, 87,
+        2, 247, 160, 250, 91, 126, 186, 47, 253, 16, 123, 88, 60, 145, 245, 3, 211, 114, 200, 12,
+        2, 134, 205, 20, 9, 21, 170, 146, 65, 40, 2, 107, 115, 225, 210, 80, 213, 137, 70, 139, 79,
+        193, 26, 18, 182, 150,
     ];
 
     assert_eq!(&EXPECTED[..], &pem_to_der(INPUT).unwrap()[..]);
@@ -215,16 +219,19 @@ impl Credentials {
 
     pub fn get_oauth_token(&self) -> Result<oauth2::Token, Error> {
         // Generate the assertion from the admin credentials
-        let assertion = create_jwt_encoded(&self,
-                                           Some(oauth2::user_scope_list().iter()),
-                                           chrono::Duration::hours(1),
-                                           Some(self.client_id.clone()),
-                                           Some(self.client_email.clone()),
-                                           oauth2::JWT_AUDIENCE_OAUTH)?;
+        let assertion = create_jwt_encoded(
+            &self,
+            Some(oauth2::user_scope_list().iter()),
+            chrono::Duration::hours(1),
+            Some(self.client_id.clone()),
+            Some(self.client_email.clone()),
+            oauth2::JWT_AUDIENCE_OAUTH,
+        )?;
 
         // Request Google Oauth2 to retrieve the access token in order to create a session cookie
         let client = reqwest::blocking::Client::new();
-        let response: oauth2::Token = client.post(oauth2::GOOGLE_OAUTH_URL)
+        let response: oauth2::Token = client
+            .post(oauth2::GOOGLE_OAUTH_URL)
             .form(&oauth2::request_form(assertion))
             .send()?
             .json()?;