+++ src/users.rs
@@ -2,20 +2,20 @@
 //!
 //! Retrieve firebase user information
 
-use std::collections::HashMap;
+use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
-use serde::de::DeserializeOwned;
+use std::collections::HashMap;
 
-use crate::FirebaseAuthBearer;
-use crate::errors::FirebaseError;
 use super::errors::{extract_google_api_error, Result};
 use super::sessions::{service_account, user};
+use crate::errors::FirebaseError;
+use crate::FirebaseAuthBearer;
 
 /// A federated services like Facebook, Github etc that the user has used to
 /// authenticated himself and that he associated with this firebase auth account.
 #[allow(non_snake_case)]
-#[derive(Debug, Default, Deserialize, Serialize)]
+#[derive(Debug, Default, Clone, Deserialize, Serialize)]
 pub struct ProviderUserInfo {
     pub providerId: String,
     pub federatedId: String,
@@ -25,7 +25,7 @@ pub struct ProviderUserInfo {
 
 /// Users id, email, display name and a few more information
 #[allow(non_snake_case)]
-#[derive(Debug, Default, Deserialize, Serialize)]
+#[derive(Debug, Default, Clone, Deserialize, Serialize)]
 pub struct FirebaseAuthUser {
     pub localId: Option<String>,
     pub email: Option<String>,
@@ -51,7 +51,7 @@ impl FirebaseAuthUser {
         let attr = self.customAttributes.clone();
         match attr {
             Some(s) => Ok(serde_json::from_str(&s)?),
-            None => Ok(T::default())
+            None => Ok(T::default()),
         }
     }
 }
@@ -63,9 +63,16 @@ pub struct FirebaseAuthUserResponse {
 }
 
 #[allow(non_snake_case)]
-#[derive(Serialize)]
-struct UserRequest {
-    pub idToken: String,
+#[derive(Serialize, Debug)]
+pub struct UserRequest {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub idToken: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub uid: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub email: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub phoneNumber: Option<String>,
 }
 
 #[allow(non_snake_case)]
@@ -98,7 +105,9 @@ impl UserUpdateRequest {
         let deserialized: HashMap<&str, Value> = serde_json::de::from_str(&serialized)?;
         for &name in reserved_claims().iter() {
             if deserialized.contains_key(name) {
-                return Err(FirebaseError::Generic("claim is reserved and must not be set"));
+                return Err(FirebaseError::Generic(
+                    "claim is reserved and must not be set",
+                ));
             }
         }
 
@@ -109,19 +118,40 @@ impl UserUpdateRequest {
 
 #[inline]
 fn firebase_auth_url(action: &str, key: &str) -> String {
-    format!("https://identitytoolkit.googleapis.com/v1/accounts:{}?key={}", action, key)
+    format!(
+        "https://identitytoolkit.googleapis.com/v1/accounts:{}?key={}",
+        action, key
+    )
 }
 
 #[inline]
 fn firebase_auth_project_url(action: &str, project: &str, key: &str) -> String {
-    format!("https://identitytoolkit.googleapis.com/v1/projects/{}/accounts:{}?key={}", project, action, key)
+    format!(
+        "https://identitytoolkit.googleapis.com/v1/projects/{}/accounts:{}?key={}",
+        project, action, key
+    )
 }
-
 
 #[inline]
 fn reserved_claims() -> [&'static str; 16] {
-    ["acr", "amr", "at_hash", "aud", "auth_time", "azp", "cnf", "c_hash",
-        "exp", "firebase", "iat", "iss", "jti", "nbf", "nonce", "sub", ]
+    [
+        "acr",
+        "amr",
+        "at_hash",
+        "aud",
+        "auth_time",
+        "azp",
+        "cnf",
+        "c_hash",
+        "exp",
+        "firebase",
+        "iat",
+        "iss",
+        "jti",
+        "nbf",
+        "nonce",
+        "sub",
+    ]
 }
 
 /// Retrieve information about the firebase auth user associated with the given user session
@@ -136,7 +166,10 @@ pub fn user_info(session: &user::Session) -> Result<FirebaseAuthUserResponse> {
         .client()
         .post(&url)
         .json(&UserRequest {
-            idToken: session.access_token(),
+            idToken: Some(session.access_token()),
+            uid: None,
+            email: None,
+            phoneNumber: None,
         })
         .send()?;
 
@@ -145,15 +178,101 @@ pub fn user_info(session: &user::Session) -> Result<FirebaseAuthUserResponse> {
     Ok(resp.json()?)
 }
 
+/// Retrieve information about the firebase auth user associated with the given uid
+///
+/// Error codes:
+/// - INVALID_ID_TOKEN
+/// - USER_NOT_FOUND
+pub fn get_user(session: &service_account::Session, uid: &str) -> Result<FirebaseAuthUser> {
+    get_user_info(
+        session,
+        UserRequest {
+            idToken: None,
+            uid: Some(uid.to_string()),
+            email: None,
+            phoneNumber: None,
+        },
+    )
+}
+
+/// Retrieve information about the firebase auth user associated with the given email
+///
+/// Error codes:
+/// - INVALID_ID_TOKEN
+/// - USER_NOT_FOUND
+pub fn get_user_by_email(
+    session: &service_account::Session,
+    email: &str,
+) -> Result<FirebaseAuthUser> {
+    get_user_info(
+        session,
+        UserRequest {
+            idToken: None,
+            email: Some(email.to_string()),
+            uid: None,
+            phoneNumber: None,
+        },
+    )
+}
+
+/// Retrieve information about the firebase auth user associated with the given email
+///
+/// Error codes:
+/// - INVALID_ID_TOKEN
+/// - USER_NOT_FOUND
+pub fn get_user_by_phone(
+    session: &service_account::Session,
+    phone: &str,
+) -> Result<FirebaseAuthUser> {
+    get_user_info(
+        session,
+        UserRequest {
+            idToken: None,
+            email: None,
+            uid: None,
+            phoneNumber: Some(phone.to_string()),
+        },
+    )
+}
+
+fn get_user_info(
+    session: &service_account::Session,
+    request: UserRequest,
+) -> Result<FirebaseAuthUser> {
+    let url = firebase_auth_url("lookup", &session.credentials.api_key);
+
+    let resp = session
+        .client()
+        .post(&url)
+        .bearer_auth(session.oauth_access_token().to_owned())
+        .json(&request)
+        .send()?;
+
+    let resp = extract_google_api_error(resp, || format!("{:?}", request))?;
+    let result: FirebaseAuthUserResponse = resp.json()?;
+    match result.users.first() {
+        Some(user) => Ok(user.clone()),
+        None => Err(FirebaseError::Generic("empty result")),
+    }
+}
+
 /// Updates the firebase auth user data associated with the given user session
 ///
 /// Error codes:
 /// - INVALID_ID_TOKEN
 /// - USER_NOT_FOUND
-pub fn user_set_claims<T: Serialize>(session: &service_account::Session, user_id: &str, claims: T) -> Result<()> {
+pub fn user_set_claims<T: Serialize>(
+    session: &service_account::Session,
+    user_id: &str,
+    claims: T,
+) -> Result<()> {
     let req = UserUpdateRequest::new(user_id.clone()).update_custom_claims(claims)?;
 
-    let url = firebase_auth_project_url("update", &session.credentials.project_id,&session.credentials.api_key);
+    let url = firebase_auth_project_url(
+        "update",
+        &session.credentials.project_id,
+        &session.credentials.api_key,
+    );
     let resp = session
         .client()
         .post(&url)
@@ -164,7 +283,6 @@ pub fn user_set_claims<T: Serialize>(session: &service_account::Session, user_id
     extract_google_api_error(resp, || user_id.to_owned())?;
     Ok({})
 }
-
 
 /// Removes the firebase auth user associated with the given user session
 ///
@@ -177,7 +295,10 @@ pub fn user_remove(session: &user::Session) -> Result<()> {
         .client()
         .post(&url)
         .json(&UserRequest {
-            idToken: session.access_token(),
+            idToken: Some(session.access_token()),
+            uid: None,
+            email: None,
+            phoneNumber: None,
         })
         .send()?;
 
@@ -188,7 +309,7 @@ pub fn user_remove(session: &user::Session) -> Result<()> {
 #[allow(non_snake_case)]
 #[derive(Default, Deserialize)]
 struct SignInUpUserResponse {
-    localId: String,
+    localId: Option<String>,
     idToken: String,
     refreshToken: String,
 }
@@ -208,7 +329,12 @@ struct SignInCustom {
     pub returnSecureToken: bool,
 }
 
-fn sign_up_in(session: &service_account::Session, email: &str, password: &str, action: &str) -> Result<user::Session> {
+fn sign_up_in(
+    session: &service_account::Session,
+    email: &str,
+    password: &str,
+    action: &str,
+) -> Result<user::Session> {
     let url = firebase_auth_url(action, &session.credentials.api_key);
     let resp = session
         .client()
@@ -226,7 +352,7 @@ fn sign_up_in(session: &service_account::Session, email: &str, password: &str, a
 
     Ok(user::Session::new(
         &session.credentials,
-        Some(&resp.localId),
+        resp.localId.as_deref(),
         Some(&resp.idToken),
         Some(&resp.refreshToken),
     )?)
@@ -239,7 +365,11 @@ fn sign_up_in(session: &service_account::Session, email: &str, password: &str, a
 /// EMAIL_EXISTS: The email address is already in use by another account.
 /// OPERATION_NOT_ALLOWED: Password sign-in is disabled for this project.
 /// TOO_MANY_ATTEMPTS_TRY_LATER: We have blocked all requests from this device due to unusual activity. Try again later.
-pub fn sign_up(session: &service_account::Session, email: &str, password: &str) -> Result<user::Session> {
+pub fn sign_up(
+    session: &service_account::Session,
+    email: &str,
+    password: &str,
+) -> Result<user::Session> {
     sign_up_in(session, email, password, "signUp")
 }
 
@@ -249,7 +379,11 @@ pub fn sign_up(session: &service_account::Session, email: &str, password: &str) 
 /// EMAIL_NOT_FOUND: There is no user record corresponding to this identifier. The user may have been deleted.
 /// INVALID_PASSWORD: The password is invalid or the user does not have a password.
 /// USER_DISABLED: The user account has been disabled by an administrator.
-pub fn sign_in(session: &service_account::Session, email: &str, password: &str) -> Result<user::Session> {
+pub fn sign_in(
+    session: &service_account::Session,
+    email: &str,
+    password: &str,
+) -> Result<user::Session> {
     sign_up_in(session, email, password, "signInWithPassword")
 }
 
@@ -259,7 +393,11 @@ pub fn sign_in(session: &service_account::Session, email: &str, password: &str) 
 /// EMAIL_NOT_FOUND: There is no user record corresponding to this identifier. The user may have been deleted.
 /// INVALID_PASSWORD: The password is invalid or the user does not have a password.
 /// USER_DISABLED: The user account has been disabled by an administrator.
-pub fn sign_in_with_custom_jwt(session: &service_account::Session, token: &str) -> Result<user::Session> {
+pub fn sign_in_with_custom_jwt(
+    session: &service_account::Session,
+    user_id: &str,
+    token: &str,
+) -> Result<user::Session> {
     let url = firebase_auth_url("signInWithCustomToken", &session.credentials.api_key);
     let resp = session
         .client()
@@ -273,11 +411,7 @@ pub fn sign_in_with_custom_jwt(session: &service_account::Session, token: &str) 
     let resp = extract_google_api_error(resp, || "signInWithCustomToken".to_owned())?;
 
     let resp: SignInUpUserResponse = resp.json()?;
-
-    Ok(user::Session::new(
-        &session.credentials,
-        Some(&resp.localId),
-        Some(&resp.idToken),
-        Some(&resp.refreshToken),
-    )?)
+    let mut session = user::Session::by_access_token(&session.credentials, &resp.idToken)?;
+    session.user_id = user_id.to_string();
+    Ok(session)
 }